requirements.txt
flet==0.23.2
sqlalchemy
pg8000
pydantic-settings
python-dotenv
pillow